feat(wallet): add WalletConnect component and update NavBar
Extract wallet connect/disconnect UI into a standalone WalletConnect
component that shows the shortened address when connected and an error
message when connection fails. Update NavBar to use WalletConnect and
add Marketplace and Dashboard nav links.

Closes #188
